requirements.txt
flask-assets==2.0 ; python_version >= "3.10" and python_version < "4.0"
flask==2.3.3 ; python_version >= "3.10" and python_version < "4.0"
flask-restful
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
werkzeug==2.3.7 ; python_version >= "3.10" and python_version < "4.0"
flask_cors
flask-login
flask-restful 
pymysql
mysql-connector-python
